Rename dovecot.cf to dovecot-services.cf
Prior to this change there was a file named `config/dovecot.cf` which appeared to be placed to chart users could add arbitrary custom Dovecot configuration. This name conflicted with the services defined in `templates/configmap.yaml`, so it was never actually staged. This PR renames the one in `configmap.yaml` so both may be used.