Document possible values for ldap_start_tls
I had this set to the value of yes (no quotes), which was interpreted as a boolean somewhere in the chain, causing my rendered configuration to have `start_tls = true`. Unfortuantely postfix only accepts a literal `yes` or `no`. Thus, to make sure it gets all the way to Postfix in the correct form, it is important that this string is quoted.